chore(fluxWizard): refactor flux generation code
